prune files
