+++ schemas/index.js
@@ -3,6 +3,7 @@ const mongoose = require('mongoose');
 const { MONGO_ID, MONGO_PASSWORD, NODE_ENV } = process.env;
 const MONGO_URL = `mongodb://${MONGO_ID}:${MONGO_PASSWORD}@localhost:27017/admin`;
 console.log(MONGO_URL);
+
 const connect = () => {
   if (NODE_ENV !== 'production') {
     mongoose.set('debug', true);
